requirements.txt
# ColdOS 防谄媚场景集成框架 - 依赖文件

# 前端框架
streamlit==1.52.2
plotly==6.5.1
pandas==2.3.3

# 大模型API调用
dashscope==1.16.1

# 配置文件处理
pyyaml==6.0.1

# 可选：CORS支持（如果需要Flask后端）
Flask==3.1.3
Flask-CORS==6.0.2